requirements.txt
Django==5.2
djangorestframework==3.16.0
djangorestframework-simplejwt==5.3.1
python-decouple==3.8
drf-spectacular==0.27.2
pytest-django==4.8.0
aiohttp==3.10.11
django-cors-headers==4.6.0
django-ratelimit==4.1.0
PyMySQL==1.1.1
django-filter==24.3

